Refactor to split each test in their individual files
